refactor: move shared code(isAbsoluteUrl) to utils and archive old files(getSubMenu).
