+++ src/features/common/Layout/Navbar/microComponents/Archive/SubMenu.tsx
@@ -1,7 +1,7 @@
 import type { SubmenuItem } from '@planet-sdk/common';
 
 import { lang_path } from '../../../../../../utils/constants/wpLanguages';
-import GetSubMenuIcons from '../../getSubMenu';
+import GetSubMenuIcons from './getSubMenu';
 import { useTranslations, useLocale } from 'next-intl';
 import Link from 'next/link';
