+++ src/features/common/Layout/Navbar/utils.ts
@@ -39,3 +39,18 @@ export const isPlanetDomain = (url: string) => {
     return false;
   }
 };
+
+/**
+ * Checks if a given URL is an absolute URL.
+ *
+ * @param url - The URL to check
+ * @returns Boolean indicating if the URL is absolute
+ *
+ * @example
+ * isAbsoluteUrl('https://api.example.com') // returns true
+ * isAbsoluteUrl('/users') // returns false
+ */
+export function isAbsoluteUrl(url: string) {
+  const pattern = /^https?:\/\//i;
+  return pattern.test(url);
+}
